+++ gbdxtools/workflow.py
@@ -21,6 +21,9 @@ class Workflow(object):
         Returns:
             An instance of the Workflow class.
         """
+        self.base_url = '%s/workflows/v1' % interface.root_url
+        self.workflows_url = '%s/workflows' % self.base_url
+
         # store a reference to the GBDX Connection
         self.gbdx_connection = interface.gbdx_connection
 
@@ -41,9 +44,8 @@ class Workflow(object):
         """
 
         # hit workflow api
-        url = 'https://geobigdata.io/workflows/v1/workflows'
         try:
-            r = self.gbdx_connection.post(url, json=workflow)
+            r = self.gbdx_connection.post(self.workflows_url, json=workflow)
             try:
                 r.raise_for_status()
             except:
@@ -65,10 +67,66 @@ class Workflow(object):
              Workflow status (str).
         """
         self.logger.debug('Get status of workflow: ' + workflow_id)
-        url = 'https://geobigdata.io/workflows/v1/workflows/' + workflow_id
+        url = '%(wf_url)s/%(wf_id)s' % {
+            'wf_url': self.workflows_url, 'wf_id': workflow_id
+        }
         r = self.gbdx_connection.get(url)
 
         return r.json()['state']
+
+    def get(self, workflow_id):
+        """Get existing workflow state and task information.
+
+         Args:
+             workflow_id (str): Workflow id.
+
+         Returns:
+             Workflow object (dict).
+        """
+        self.logger.debug('Get workflow object: ' + workflow_id)
+        url = '%(wf_url)s/%(wf_id)s' % {
+            'wf_url': self.workflows_url, 'wf_id': workflow_id
+        }
+        r = self.gbdx_connection.get(url)
+        r.raise_for_status()
+
+        return r.json()
+
+    def get_stdout(self, workflow_id, task_id):
+        """Get stdout for a particular task.
+
+         Args:
+             workflow_id (str): Workflow id.
+             task_id (str): Task id.
+
+         Returns:
+             Stdout of the task (string).
+        """
+        url = '%(wf_url)s/%(wf_id)s/tasks/%(task_id)s/stdout' % {
+            'wf_url': self.workflows_url, 'wf_id': workflow_id, 'task_id': task_id
+        }
+        r = self.gbdx_connection.get(url)
+        r.raise_for_status()
+
+        return r.text
+
+    def get_stderr(self, workflow_id, task_id):
+        """Get stderr for a particular task.
+
+         Args:
+             workflow_id (str): Workflow id.
+             task_id (str): Task id.
+
+         Returns:
+             Stderr of the task (string).
+        """
+        url = '%(wf_url)s/%(wf_id)s/tasks/%(task_id)s/stderr' % {
+            'wf_url': self.workflows_url, 'wf_id': workflow_id, 'task_id': task_id
+        }
+        r = self.gbdx_connection.get(url)
+        r.raise_for_status()
+
+        return r.text
 
     def events(self, workflow_id):
         '''Get workflow events.
@@ -80,7 +138,9 @@ class Workflow(object):
              List of workflow events.
         '''
         self.logger.debug('Get events of workflow: ' + workflow_id)
-        url = 'https://geobigdata.io/workflows/v1/workflows/' + workflow_id + '/events'
+        url = '%(wf_url)s/%(wf_id)s/events' % {
+            'wf_url': self.workflows_url, 'wf_id': workflow_id
+        }
         r = self.gbdx_connection.get(url)
 
         return r.json()['Events']
@@ -95,7 +155,9 @@ class Workflow(object):
                Nothing
         """
         self.logger.debug('Canceling workflow: ' + workflow_id)
-        url = 'https://geobigdata.io/workflows/v1/workflows/' + workflow_id + '/cancel'
+        url = '%(wf_url)s/%(wf_id)s/cancel' % {
+            'wf_url': self.workflows_url, 'wf_id': workflow_id
+        }
         r = self.gbdx_connection.post(url, data='')
         r.raise_for_status()
 
@@ -110,7 +172,9 @@ class Workflow(object):
         """
 
         # hit workflow api
-        url = 'https://geobigdata.io/workflows/v1/batch_workflows'
+        url = '%(base_url)s/batch_workflows' % {
+            'base_url': self.base_url
+        }
         try:
             r = self.gbdx_connection.post(url, json=batch_workflow)
             batch_workflow_id = r.json()['batch_workflow_id']
@@ -128,7 +192,9 @@ class Workflow(object):
              Batch Workflow status (str).
         """
         self.logger.debug('Get status of batch workflow: ' + batch_workflow_id)
-        url = 'https://geobigdata.io/workflows/v1/batch_workflows/' + batch_workflow_id
+        url = '%(base_url)s/batch_workflows/%(batch_id)s' % {
+            'base_url': self.base_url, 'batch_id': batch_workflow_id
+        }
         r = self.gbdx_connection.get(url)
 
         return r.json()
@@ -143,7 +209,9 @@ class Workflow(object):
              Batch Workflow status (str).
         """
         self.logger.debug('Cancel batch workflow: ' + batch_workflow_id)
-        url = 'https://geobigdata.io/workflows/v1/batch_workflows/{0}/cancel'.format(batch_workflow_id)
+        url = '%(base_url)s/batch_workflows/%(batch_id)s/cancel' % {
+            'base_url': self.base_url, 'batch_id': batch_workflow_id
+        }
         r = self.gbdx_connection.post(url)
 
         return r.json()